Remove the recently-added orderBy in angular-tree-control
Pass-through support for the orderBy filter was added to the main
project in wix/angular-tree-control#15. However, this was breaking the
sorting for our app where we sort on our own beforehand in the backend,
and manually sort after adding new branches and leaves to the tree. It
seems like having an empty orderBy expression in the template was
leading to issues since each one of our nodes are actually objects, so a
naive sort implementation wouldn't be able to sort it properly.

This is just a temporary fix while we attempt to address the issue
upstream.