Move code mirror themes to the components section
Due to the way we're applying the theme, we need to do some extra
finagling to the themes themselves to get them to work properly. That
finagling gets overwritten when we update CodeMirror, so we should just
keep a copy of the themes ourselves that don't get overwritten.

Alternatively, we should stop finagling things, but eh we're lazy!