Fix browser/server initialization
The `ng-init` was always running even when we had a stored value, and it
was overwriting/ignoring our value from storage. We do a "simple"
ternary check (that should probably be done in the js) to handle the
null case (coming to a new session) and the existing case (refreshing
the page).