Fix #13: Prefix the server to the browser
Jenn noticed that rerunning set reports was always running locally, as
the reports were being written with their browser being 'Chrome Local'
instead of 'GP Chrome Local' for example. We have all the information
available, so we just need to prefix the server to the browser before
sending it through.