Fix odd unit test error by using an empty config
We were getting

    Script Error

in our Karma unit tests in PhantomJS, but no discernable explanation as
to why. The error was intermittent, and not always on the same test
number - sometimes it failed out on test 111, other times it would be
128. Whenever we ran a subset of the tests, we'd always pass.

We switched the test-running browser to Chrome so that we could see the
console and see the error in all its glory, which ended up finally
leading us to this solution.

The tests apparently had been sending a bunch of requests to Pusher,
because I wasn't properly mocking out the services. So, anything that
depended on the LiveReport service (which depends on the Pusher module)
would be making a few requests to Pusher.com. Oddly though, our tests on
travis were passing just fine.

The difference was that locally, we're using a fully populated
app/config.js so that the app is usable. So that means our tests were
also using a real app/config.js, which means our tests got to talk over
the wire all the way to Pusher. Meanwhile, on travis, the config
is actually empty since it's empty in the repo, and so travis wasn't
DOSing Pusher at all.

The solution was to use an empty config file specifically for tests that
would prevent us from getting hung up on Pusher errors.