Stop instantiating monitor controller twice
This was causing all http requests to be handled twice, once by each
controller instance. The bug arose when we moved from the old
router/controller strategy (specify the template and the controller in
the route) to the new one (specify the template and controller in the
route, AND specify the controller in the partial via the controllerAs
syntax.) Removing the declaration in the router resolves the issue as
the controller is only declared once and thus is only instantiating
once.