Stop persisting the trees to localStorage
We hypothesize that this is a significant performance hit - ngStorage
shows up pretty hot in the performance profiling, particular its
invocations of angular.copy, so we are testing the theory that avoiding
usage of ngStorage when possible helps performance.

Additionally, I can't recall what this is supposed to be helping us
do. I believe the intent was to cache the tree results to allow us to
show a tree immediately on pageload, but when using a fake delay
locally (sleep(5); in the PHP rest API), the tree would take 5 seconds
to display before appearing anyway. So, I don't think this code was
accomplishing anything.