Determine which filetree tab is active manually
If we don't do it on our own, it defaults to the first tab, and then it
calls getTreeContents on the features tab, even though we might not want
the features tab. Doing it this way seems to be early enough that when
the treecontrol directive is compiled, the active tab is already the tab
as determined by the path.