Fix the date shenanigans with feedback from production
For whatever reason, the production system does its dates differently
from our dev machine, so it wasn't possible to be sure our previous
effort in 57c443b actually worked without deploying it. Now that it's
deployed, we know what we should've done the first time around, and
that's what's contained in this fix.