Stop changing browser based on hostname
This was implemented when the hostname was sometimes a URL to a native
mobile app. We are no longer doing native mobile app development
locally, so we do not need to change the browser based on hostname.

In fact, this ends up being detrimental because one of our browsers is
"iOS Mobile Safari", and choosing that browser limits the potential
servers. However, if one chooses the browser first, and then chooses the
hostname after, the browser will change unexpectedly.