Replace shell deployBack with rsync version
The shell deployback was depending on the destination folder being a git
repository. Our set up script in honeydew-install don't seem to be
setting that up properly, and there's not really any need for it to be a
fully fledged git repository anyway, so we don't need to (and shouldn't)
assume that it is.